pyproject.toml
[tool.poetry]
name = "sc4001-value-alignment-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
transformers = "^4.46.2"
openai = "^1.54.3"
torch = "^2.5.1"
pandas = "^2.2.3"
parascore = "^1.0.5"
nltk = "^3.9.1"
bert-score = "^0.3.13"
rouge-score = "^0.1.2"
sentence-transformers = "^3.3.0"
protobuf = "^5.28.3"
tiktoken = "0.7.0"
blobfile = "^3.0.0"
peft = "^0.13.2"
sentencepiece = "^0.2.0"
python-dotenv = "^1.0.1"
sage = {path = "SaGE/library"}
trl = "^0.12.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
